+++ examples/simple.rb
@@ -1,7 +1,16 @@
-require "../lib/esysPinger.rb"
+#!/usr/bin/env ruby
+# coding: utf-8
+
+$:.unshift File.expand_path '../lib', File.dirname(__FILE__)
+require 'esys_pinger'
 require "pp"
 
-room = PCroom.new(2..91,timeout:5) # 検索対象のPCナンバー
-pp room.get_status # 全マシンのステータスを配列で取得
+# 検索対象のPCナンバーを指定
+room = EsysPinger::PCroom.new(2..91,timeout:5)
+
+# 全マシンのステータスを配列で取得
+pp room.get_status
+
+# それぞれの状態のマシンをカウント
 puts "on #{room.count(:on)}"
 puts "off #{room.count(:off)}"
