+++ lib/esys_pinger.rb
@@ -1,0 +1,10 @@
+# coding: utf-8
+$:.unshift File.dirname(__FILE__)
+
+require 'net/ping'
+require 'net/ssh'
+
+module EsysPinger
+  require_relative "esys_pinger/pc_node"
+  require_relative "esys_pinger/pc_room"
+end
