+++ lib/esys_pinger/pc_node.rb
@@ -1,0 +1,71 @@
+# coding: utf-8
+
+module EsysPinger
+  class PCnode
+    attr_accessor :node_num, :addr, :node_num_str,:ssh,:status
+    def initialize(num,timeout:1,ssh:nil)
+      @node_num = num
+      @ssh      = ssh
+
+      case @node_num.to_s.size
+        when 1
+          @node_num_str = '00'+ @node_num.to_s
+        when 2
+          @node_num_str = '0' + @node_num.to_s
+        else
+          raise SyntaxError.new('invalid node_num')
+      end
+      @addr = "esys-pc#{@node_num_str}.edu.esys.tsukuba.ac.jp"
+      @pinger = Net::Ping::External.new(@addr,nil,timeout)
+      @ssh[:opt][:timeout]=timeout if ssh
+    end
+
+    def on?
+      return @pinger.ping?
+    end
+
+    def linux?
+      if @ssh.nil?
+        raise SyntaxError.new('no ssh option')
+      end
+
+      begin
+        return true if Net::SSH.start(@addr,@ssh[:username],@ssh[:opt])
+      rescue
+        return false
+      end
+    end
+
+    def windows?
+      return !self.linux? && self.on?
+    end
+
+    def user
+      if !self.linux?
+        raise SyntaxError.new('not linux')
+      end
+
+      begin
+        users = ""
+        Net::SSH.start(@addr,ssh[:username],ssh[:opt]) do |s|
+          users = s.exec! 'users'
+        end
+        return users.split(" ").first
+      rescue
+        return nil
+      end
+    end
+
+    def get_status
+      #status = :Linux :windows :off
+      if @ssh
+        @status ||= :linux if self.linux?
+        @status ||= :windows if self.on?
+      else
+        @status ||= :on if self.on?
+      end
+      @status ||= :off
+      return @status
+    end
+  end
+end
